Random bugfix

+++ Editor/Extensions/Random.cs
@@ -3,10 +3,10 @@ namespace Enca.Extensions
     public static class Random
     {
         /// <summary>
-        ///     Get random number from 0 to given number.
+        /// Get random number from 0 to given number.
         /// </summary>
         /// <param name="limitNumber"></param>
-        public static float GetRandomNumber(float limitNumber) => UnityEngine.Random.Range(0, limitNumber);
+        public static int GetRandomNumber(int limitNumber) => UnityEngine.Random.Range(0, limitNumber);
         
         /// <summary>
         /// Returns a random boolean value.
